Completions: Move tactics to shared `json` file.
This unifies the two `json` files that are used for completions, in
the autocomplete within the editor as well as the tactics panel.
